[nova] switch to projected volumes
We still have crashlooping pods when the container exits
(e.g. when the process is killed manually or by exception).
Fabian Ruff found a bug, that should explain the behavior:
  https://github.com/kubernetes/kubernetes/issues/68211#issuecomment-467032048

This bug also contains a workaround, that's implemented here: use a
projected configMap instead of mounting directly from the original ones.
This way, an updated configMap will not be seen by the restarting
container and thus will not lead to the crashloop we've seen.

CCM-9905