Label all vcenter objects by vcenter/datacenter/cluster
This should allow us to more easily associated kubernetes resources with vcenter resources.
Now we can decommision a cluster or a vcenter and remove all corresponding resources in kubernetes.
In prometheus, we can associate the alert with the state of the object in netbox, disable the alerting
of resources still in build-up

Following the convention of the vrops and vcsa exporter,
vccluster is used as a label to avoid collision with kubernetes cluster,.
and the fqdn is used for vcenter.